Saiba/requirements.txt
Django==1.10.1
djangorestframework==3.4.7
Pillow==3.3.1
six==1.10.0
textile==2.3.3
